feat: navbar for checkout pages
