+++ app/components/CheckoutFormShipping.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+const formData = useCheckoutUserFormData()
 </script>
 
 <template>
