+++ app/components/HeroEffect.vue
@@ -28,7 +28,10 @@ onBeforeUnmount(() => {
   <div class="absolute inset-0 -z-10 transform-gpu blur-3xl">
     <div
       ref="path"
-      class="size-full bg-primary transition-[clip-path] ease-in-out duration-(--hg-hero-effect-animation-duration)"
+      class="
+        size-full bg-primary transition-[clip-path]
+        duration-(--hg-hero-effect-animation-duration) ease-in-out
+      "
       :style="`--hg-hero-effect-animation-duration: ${duration}ms;`"
       style="clip-path: polygon(99.401% 23.7078%, 35.7645% 92.5042%, 56.7894% 30.1775%, 21.3293% 9.14554%, 85.9612% 55.7237%, 91.7772% 28.9103%, 68.0925% 83.52%, 80.452% 82.1599%, 40.6311% 70.1134%, 74.8706% 37.9842%, 59.4388% 31.0837%, 19.5192% 72.8712%, 52.9892% 58.0796%, 64.8176% 97.7889%, 101.755% 95.5604%, 95.0635% 4.17111%);"
     />
