+++ app/components/NavbarCheckout.vue
@@ -1,12 +1,14 @@
 <template>
-  <NuxtLink to="/" class="group">
-    <SvgLogoHeartlessgamingSkull2020
-      class="
-        box-content size-12 p-4 transition-transform
-        group-hover:-rotate-12
-        sm:p-6
-      "
-      :font-controlled="false"
-    />
-  </NuxtLink>
+  <div class="flex items-center justify-between">
+    <NuxtLink to="/" class="group">
+      <SvgLogoHeartlessgamingSkull2020
+        class="
+          box-content size-12 p-4 transition-transform
+          group-hover:-rotate-12
+          sm:p-6
+        "
+        :font-controlled="false"
+      />
+    </NuxtLink>
+  </div>
 </template>
