+++ app/utils/toCurrency.ts
@@ -1,4 +1,6 @@
-export default function toCurrency(value: number | string, lang = 'fr-FR', currency = 'EUR') {
+type Value = number | string
+
+export default function toCurrency(value: Value, lang = 'fr-FR', currency = 'EUR') {
   return new Intl.NumberFormat(lang, { style: 'currency', currency }).format(
     value,
   )
