+++ server/api/place-order.ts
@@ -1,5 +1,5 @@
 export default defineEventHandler(async (event) => {
   // make a stripe payment
-  
+
   return { topkek: true }
 })
